Fix TBot crashing on invalid Unicode
Some boards (was noticed with p2020rdb) send some weird
bytes on startup. Previously, TBot would crash when this
happened. Now, some Debug information will be printed to
stdout and TBot will attempt to decode the data as "latin_1".
If this fails as well, TBot will crash as previously.
This is a temporary solution and will be fixed properly
at some point.