Change bound var name in with statements to tb
Previously, all TBot code looked like this:

with tb.machine(...) as tbn:
    ...

This had the issue of not shadowing the tb variable and thus
allowing a programmer to make mistakes by using tb when
tbn was meant. By changing to

with tb.machine(...) as tb:
    ...

this is no longer possible.