Fix tb leaking from with-statement
When using a construct like this:

tb1 = tb
with tb.machine(...) as tb:
    tb2 = tb
tb3 = tb

tb3 is not tb1 but tb2. This is counterintuitive and
is fixed in this commit. However, python's "is" will
still not report it fixed, because an object can't
change a reference to it. "tb1.machines is tb3.machines"
will be True with this commit if you do need that check.

Signed-off-by: Rahix <hws@denx.de>