tbot:  Add named_testcase decorator
Two testcases with the same name in different modules
(e.g. uboot.build and linux.build) would have the same
name in the log-file, which would make writing backends
significantly harder.

This commit adds a new decorator called `named_testcase`,
which allows defining a testcase with a different name
for logging.  This also affects the name of the testcase
on the commandline (you will have to use the new name now).

To pick up the above example, you would define uboot.build
as follows:

    @tbot.named_testcase("uboot_build")
    def build(*args):
        ...

Other testcases can still call it like:

    from tbot.tc.uboot import build

    ...
    build()

But in the log-file and for calling it from the commandline
it is now called "uboot_build".

Signed-off-by: Harald Seiler <hws@denx.de>