tc.uboot: Rewrite build testcase
This commit contains a complete rewrite of the U-Boot build testcase.
The new system aims to be more flexible and allows using this test
under even more circumstances.  For example you can now easily
use it to build U-Boot while bisecting.  Even if you have to apply
patches!

Please consult the changelog for a migration guide.

Signed-off-by: Harald Seiler <hws@denx.de>