board.linux: Remove trailing whitespace from login_prompt
It may happend that on-board user space SW floods the console to
which tbot is connected.
In this situation the login_prompt read from the channel gets
malformed and the last whitespace char is missing.

In such a situation we would not regard the "login: " as correct
match.

Reduce the probability of this issue by removing the trailing space
- tbot would only look for "login:" for prompt login match.

Signed-off-by: lukma <lukma@denx.de>