channel.subprocess: Fix pty not being closed
tbot erroneously did not close the pty when closing the channel,
which allowed processes that were still running to continue lingering
until tbot exited.  This lead to weird behaviour like picocom not
being able to reacquire the console.

Just closing the pty is not enough, though.  tbot also has to wait for
all session processes to actually terminate.  To do this, this commit
introduces a check implemented using exponential backoff that gives
processes up to 1 second time to complete.  If they don't, tbot emits
a warning and tries to continue anyway.

Signed-off-by: Harald Seiler <hws@denx.de>