path: Fix __fspath__ returning wrong info
The __fspath__() method defined by the os.PathLike interface is
assumed to return a representation of the path on the local file-
system.  As tbot-paths are potentially non-local the existence of
this method does not make sense in the first place anyway.
    However, as tbot's Path type is derived from pathlib.Path, it
inherits this method which then returns the str(path) representation.
    This commit ensures that any attempt to use __fspath__() will
result in an exception, to make certain that it is not possible to
introduce silent bugs by functions from the python standard library
accepting a tbot-path.

Signed-off-by: Harald Seiler <hws@denx.de>