machine-v2: Add LinuxShell.open_channel()
open_channel() mimicks the old LabHost.new_channel().  An important
distinction is that open_channel() consumes the current machine so it is
no longer accessible afterwards.  This is an intentional decision.  The
pattern to use for a new channel is:

        with some_machine.clone() as cl:
            chan = cl.open_channel("telnet", "192.0.2.1")

This ensures proper cleanup, in contrast to the old implementation.  It
also allows more flexible creation of channels which lie further down
the init-chain, for example behind multiple SSH jumps.

Signed-off-by: Harald Seiler <hws@denx.de>