machine-v2: Revert back to separated board design
Initially I thought it would make sense to get rid of the split design
where one has to init the board first and the U-Boot.  Implementing
this, I noticed how it leads to difficulties in modelling the
interaction between U-Boot and Linux.

The new design is again split into a `Board` and a `U-Boot`.  This time,
however they are both equivalent objects: Both are machines.  The
`Board` is a `RawShell` and will in most cases be defined as:

        class MyBoard(
            connector.ConsoleConnector,
            board.PowerControl,
            board.Board,
        ):
            name = "foo"

The board allows raw access to the serial console.  It is instanciated
by cloning a connection to the LabHost, ie:

        with MyBoard(lh) as b:
            ...

Next, the U-Boot is a `UBootShell` and uses the new `board.Connector`.
This connector just borrows the Board's channel for the U-Boot machine:

        class MyUBoot(
            board.Connector,
            [board.UBootAutobootIntercept,]
            board.UBootShell,
        ):
            name = "foo-uboot"

And it is used like this:

        with MyBoard(lh) as b, MyUBoot(b) as ub:
            ...

Signed-off-by: Harald Seiler <hws@denx.de>