bash: Suppress loading of profile scripts
Some distros (ab)use the profile script to configure the shell.  This is
usually done in the bashrc which was already suppressed using --norc.
This patch also suppresses the loading of profiles scripts with
--noprofile which should make tbot even more stable on those systems.

It remains to be seen whether missing profile will lead to issues but as
far as I can tell it should not; the environment will always have been
set for the spawned shell's parent process already and should thus be
inherited without issues.

Signed-off-by: Harald Seiler <hws@denx.de>