bash: Mangle prompt when setting PS1
If the shell for some reason takes longer to get back to the prompt
after tbot initially detecting it, the sent command will appear twice on
the channel which can confuse tbot.  This can be prevented by mangling
the prompt in the sent command and thus avoiding any possibility of tbot
detecting the prompt before it should.

Signed-off-by: Harald Seiler <hws@denx.de>