ssh-machine: Allow initialization without a host
In some cases, the paramiko-connector is not powerful enough to connect
to the lab-host  (For example if special SSH configs are needed).  But
the SSHConnector cannot be used for these cases because it needs a
pre-existing connection.  Instead of requiring this, allow not passing
anything and make the SSHConnector instanciate a local machine to use as
its basis.  This makes the SSHConnector usable for a lab-host.

Reported-by: Claudius Heine <ch@denx.de>
Signed-off-by: Harald Seiler <hws@denx.de>