tc.shell: Allow copying from local to ssh-machine
When using SSHConnector for the lab-host, copying to and from it should
be supported.  Introducing this might lead to some issues when trying to
copy from an ssh-machine which was configured to be "behind" another
connection but we cannot prevent that properly anyway.

Add support for the following two scenarios:

 - From *ssh-machine* to *local-machine*
 - From *local-machine* to *ssh-machine*

Reported-by: Claudius Heine <ch@denx.de>
Signed-off-by: Harald Seiler <hws@denx.de>