tc.shell: Add check_for_tool testcase
When writing testcases as generic as possible, one often has to account
for different hosts having different tools installed.  Add the
check_for_tool() testcase which allows comfortable checking for the
availability of a certain tools.

check_for_tool() caches the results so it does not unnecessarily clutter
the log-file with command invocations and testcase calls.

Signed-off-by: Harald Seiler <hws@denx.de>