selftest: Fix a `sleep` still sleeping after test end
Kill a `sleep` subprocess in the selftests to ensure it does not live
past the test end where it would need to be killed by the session
ending.  Ideally, the subprocess channel would be able to properly kill
all child processes but this far I was not able to come up with a 100%
watertight solution.  By preventing this sleep from sleeping too long,
we will at least not have the selftests failing because of it.

Reported-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Harald Seiler <hws@denx.de>