linux-shell: Implement interactive commands
Add a new method LinuxShell.run() which allows spawning a command and
interacting with it.  This is implemented via a context-manager inside
which you can use the methods of `Channel` to talk to the running program.
It is designed to robustly detect as many error conditions as possible,
but by the nature in which tbot interacts with machines, it cannot be
guaranteed to always error out early on failure.

Once a testcase is done interacting with a command it MUST call either
.terminate0() or .terminate() to ensure proper cleanup.  terminate0() is
analogous to exec0() and will raise an exception if the command does not
exit with return code 0.  Both of these methods will return remaining
output of the command.

As an example, the following testcase interacts with GDB to gain some
insight into a program:

    import tbot

    @tbot.testcase
    @tbot.with_lab
    def test_gdb(lh):
        exe = lh.fsroot / "/tmp/main"

        with lh.run("gdb", "-n", exe) as gdb:
            # Interact with gdb in this context
            # Wait for gdb to start up
            gdb.read_until_prompt("(gdb) ")

            # Better for automated interaction
            gdb.sendline("set confirm off")
            gdb.read_until_prompt("(gdb) ")

            # Necessary so output is not clobbered with escape sequences
            gdb.sendline("set style enabled off")
            gdb.read_until_prompt("(gdb) ")

            gdb.sendline("break main")
            gdb.read_until_prompt("(gdb) ")

            gdb.sendline("run")
            # We have hit the breakpoint
            gdb.read_until_prompt("(gdb) ")

            gdb.sendline("info locals", read_back=True)
            local_info = gdb.read_until_prompt("(gdb) ").strip()

            for line in local_info.split("\n"):
                var, val = line.split(" = ", 1)
                tbot.log.message(f"Local variable `{var}` has value `{val}`!")

            gdb.sendline("quit")
            gdb.terminate0()

        lh.exec0("size", exe)

Signed-off-by: Harald Seiler <hws@denx.de>