linux-shell: Allow spawning subshells with custom commands
It is often useful to be able to spawn a subshell with a modified
commandline.  This allows, for example, to elevate privileges with sudo:

    with lnx.subshell("sudo", "-ni"):
        # Now as root!
        ...

There is one important restriction, however: The subshell **must** use the
same underlying shell as the outer one.  Spawning an arbitrary other shell
is not (yet) supported.

Signed-off-by: Harald Seiler <hws@denx.de>