tc.uboot: Add integration with U-Boot's test/py
Add a testcase to call U-Boot's test/py test-framework and run it against the
currently selected board in tbot.

This turns out to be quite tricky because we want to run test/py on the
build-host and have it talk to the configured board which is potentially
connected somewhere completely else.  To make this work, tbot plays proxy
between the two and mediates the serial console IO + test/py's requests to
powercycle the board.  This proxying is done using a few named pipes on the
build-host and lots of cloned build-host machines.

The testcase also has support for passing in a boardenv.  If none is given
explicitly it probes the board-config for one (`testpy_boardenv` attribute of
the `UBootBuilder`).

Signed-off-by: Harald Seiler <hws@denx.de>