selftest: Fix uboot_noab test when bash is slow
When bash takes a while to start up, the uboot_noab testcase will fail
because it aborts bash loading the fake "U-Boot" environement.  This
will then make tbot hang because it infinitely waits for the prompt.

By delaying the startup until the first prompt appears we can ensure
that tbot will never interfere with bash startup even if it is slow.

Signed-off-by: Harald Seiler <hws@denx.de>