Make tbot.testcase also work as context-manager
Allow using the `tbot.testcase` function not only as a decorator, but
also as a context-manager to mark a section as a separate test.  This
can in the future replace a common pattern of defining an ad-hoc
function to create sub-tests.

An additional benefit is that this change allows deduplicating the
decorate implementations of `tbot.testcase` and `tbot.named_testcase`.

In code use of the new context-manager functionality will look like this:

    with tbot.testcase("test_name"):
        ...

Calling `tbot.skip()` inside this block will jump to the end of the block
with no further side effects.

Signed-off-by: Harald Seiler <hws@denx.de>