bash: Fix m.env() eating some env-var values
m.env() currently does not properly return the value of environment
variables if they look like options to echo.  For example:

    # Setting works as expected
    m.env("TEST_VAR", "-E")

    # Reading returns the wrong result
    val = m.env("TEST_VAR")

`val` will be an empty string because echo interpreted the variable's
value as an option.

To fix this, change the command which is run to retrieve to value to

    echo " ${VAR}"

The space in front of the expansion ensures that echo will not interpret
any value as an option.  It is later removed by string-slicing to ensure
downstream code does not notice this hack.

This bug was found thanks to fuzzing with hypothesis:

    https://hypothesis.readthedocs.io

Signed-off-by: Harald Seiler <hws@denx.de>