u-boot: Make env-var reading more robust
Reading an environment variable using the old method could sometimes
fail.  This is because when running `echo "${var}"`, some shell-quoting
rules still apply in U-Boot and variables containing special characters
would lead to issues.  For example, the following calls behave
incorrectly:

    ub.env("test_var", "\\")
    ub.env("test_var", "'")

To fix this, switch to `printenv` instead of `echo` which prints the
values literally.  This is the more robust approach.

Signed-off-by: Harald Seiler <hws@denx.de>