contrib: Add initial (empty) module
Add a `tbot_contrib` module which is meant to be used as a place for
testcases that don't quite belong into tbot core but might be useful for
tbot users so they should be part of the main repo.  It is empty as of
now but will (hopefully!) be filled with testcases useful for
a multitude of different things over time.

If you are using tbot in `develop` mode, you need to reinstall it for this
module to become available:

    pip3 uninstall tbot
    python3 setup.py develop --user

Signed-off-by: Harald Seiler <hws@denx.de>