channel: Show a channel identifier in debug output
When tbot runs with -vvv, all data sent/received via all channels is
printed to stdout.  This is a useful debugging feature if testcases
don't quite do what they should.  However, it is not easy to see which
channel a send/receive happened on / distinguish between the different
channels open at the same time.

Add an identifier to each debug output line that helps separate which
channel a send/receive happened on.

Signed-off-by: Harald Seiler <hws@denx.de>