contrib: utils: Add testcase to find IP address of machine
In times where most hosts have many IP addresses, this is not as trivial
as one would like.  This testcase approaches the problem by trying to
find the IP-address, the host would use on a certain route.

By default, this is the route to reach a theoretical host at 1.0.0.0.
This will yield a sensible result in *most* cases but of course will not
always be the address you want.  For more fine-grained control you can
pass a `route_target`.  This is the IP-address of this theoretical host
to reach.

Signed-off-by: Harald Seiler <hws@denx.de>