Support creating a patch file with local changes
It is handy to be able to test the current U-Boot source on a board. At
present you have to create a commit first. Add support for making use of
a local patch file and applying it to the source before building.

With this it is possible to write a script which builds and tests the
current source for you.

Signed-off-by: Simon Glass <sjg@chromium.org>