Change the meaning of the 'clean' flag
At present tbot does nothing with the source unless 'clean' is provided.
This makes it impossible to do incremental builds, thus making everything
take much longer.

For quick testing it is useful to switch commands, configure and build.
Update the logic to support this. It is still possible to use the 'clean'
flag when a clean build is required.

Signed-off-by: Simon Glass <sjg@chromium.org>