test/py: Issue a warning for incorrectly configured lab-hosts
It seems to be a common pattern for people to configure their lab-host
like this:

    def build(self):
        return self

This is okay most of the time but breaks horribly in the uboot_testpy
testcase: Here, commands are potentially run in parallel on the lab-host
and build-host.  If the two are one and the same machine, this of course
can't work.

As a workaround, warn when detecting this case and attempt to fix it by
calling clone() manually.

Signed-off-by: Harald Seiler <hws@denx.de>