connector: Add from_context() classmethod
Add a from_context() classmethod to connectors (and with that, to
machines) which defines policy how a certain machine can be instanciated
from pre-requisites in a context.

In more practical terms, e.g. a ConsoleConnector will first request the
lab-host from the context before using this instance to connect itself.

from_context() is a context-manager that yield a **fully initialized**
machine instance, as detailed in the method documentation.  This means
that entering the from_context() context should also enter the machine
instance's context.

Signed-off-by: Harald Seiler <hws@denx.de>