tbot: Add tbot.role module
Roles are the "interface" between the config and testcases:  The
configuration registers machines for certain roles and testcases then
request instances of a machine for a role.

tbot comes with a set of default roles that testcases can rely on.  But
for complex setups, defining custom roles is possible just as well!

The tbot.role module also takes care of registering a few default
machines for tbot.role.LabHost, tbot.role.LocalHost, and a compatibility
proxy for tbot.role.BuildHost.

Signed-off-by: Harald Seiler <hws@denx.de>