context: Allow exclusive instance requests
Provide an exclusive argument to the request() method for instance
managers which, in contrast to normal instance requesting, ensures that
no other (later) request can get access to the instance at the same
time.

After releasing the exclusively requested instance, it will be
deinitialized so following requests will need to re-initialize it.

This commit does not (yet) change the public API surface.

Signed-off-by: Harald Seiler <hws@denx.de>