context: Allow adding custom contexts to ContextHandle
The `ContextHandle` proxy is meant to be used in places where one would
usually go with contextlib's `ExitStack`.  However, using the more
convenient `ContextHandle` is only possible when no 'non-machine'
contexts are attached.

Expose an `enter_context()` method on `ContextHandle` which behaves
exactly like the method of the same name on `ExitStack`.  This way, even
more complex scenarios can make use of `ContextHandle`.

Signed-off-by: Harald Seiler <hws@denx.de>