tc: Use context API in interactive_* testcases
Some new lab- and board-configs might use features that are only
available via the context mechanism.  To make the interactive_*
testcases work for them as well, update to the new API.

This will be 100% compatible with the old configuration system
(tbot.selectable) but also work with newer, possibly
backwards-incompatible configs.

Signed-off-by: Harald Seiler <hws@denx.de>