connector: ssh: Fix contexts breaking ssh lab-host usecase
The SSHConnector supports two modes of operation.  It can either connect
from the localhost to a remote (if no argument is passed to the
constructor) or if can connect from the lab-host to a remove (if
a lab-host instance is passed to the constructor).

The from_context() implementation added in commit 567a75a59434
("treewide: Implement from_context() for most connectors") only really
implements the latter mode.  Trying to invoke the former (e.g. by making
the lab-host use an SSHConnector) will lead to an infinite recursion as
the from_context() implementation will recursively try to request an
instance of its own class (the lab-host).

As a workaround, enable this one specific usecase again by checking if
the class the SSHConnector is initializing for is registered as the
lab-host and if so, not re-requesting itself.

In the future, it might be a better option to introduce a way to
configure which role the SSHConnector tries to use as the base for its
connection.

Fixes: 567a75a59434 ("treewide: Implement from_context() for most connectors")
Signed-off-by: Harald Seiler <hws@denx.de>