context: Add keep_alive mode
Add a configurable mode which changes the behavior of the Context:
Instead of de-initializing an instance once the last requesting
context-manager exits, the instance is "kept alive" so it can be
immediately reaquired by a following request, without
a re-initialization.  To demonstrate this, consider the following code:

    with ctx.request(tbot.role.LabHost) as lh:
        lh.env("TEST_VAR", "1234")

    with ctx.request(tbot.role.LabHost) as lh2:
        tbot.log.message(lh.env("TEST_VAR"))

With the default and previous behavior, the message would be empty.
With keep-alive mode enabled, the message would print `1234`.

This is meant to allow testcases running in sequence to re-use previous
instances instead of initializing their own each time.

This new mode can be enabled by passing `keep_alive=True` to the
constructor of `Context`.

Signed-off-by: Harald Seiler <hws@denx.de>