context: Add reset_on_error request configuration
In some situations, it makes sense to not keep an instance around if an
exception occured while it was active.  Especially with the new
`keep_alive` Context mode, this ensure we won't leave dirty/broken
instances alive.

Introduce a new `reset_on_error` configurable parameter to the request()
function which, when enabled, instructs the Context to tear down an
instance if the context-manager was exited abnormally with an exception.

Signed-off-by: Harald Seiler <hws@denx.de>