selftest: Add framework for pytest based selftests
This is the first step in migrating the selftests to run with pytest
instead of a handrolled test framework.

Add a minimal test configuration which exposes a few test-machines which
are specifically designed for selftests.  A `tbot_context` fixture is
provided which makes a tbot Context available to tests.  This Context is
configured to keep instances alive and to reset them on errors.  This
helps speeding up test execution as much as possible.

Additionally, a fixture is provided which allows running certain tests
against all supported Linux shells.

Signed-off-by: Harald Seiler <hws@denx.de>