contrib: Add machine locking implementation
Added machine locking implementation: A LockManagerBase class is
specified from which individual LockManagers can then be derived,
which can then be inherited into the Labhost. A default LockManager
based on lock files and flock(1) is supplied. Furthermore, the
classes MachineLock and PooledMachineLock are supplied which can
be inherited into the Board Machine class, to request individual
locks, or locks from pools, respectively.

Signed-off-by: Niel Fourie <lusus@denx.de>