selftest: Implement mock-SSH tests using pytest
Previously, we had the `minisshd` utility to "simulate" connection to
a remote host via SSH.  This was based on dropbear which meant it
wouldn't run for most people.

Deprecate the approach in favor of one based on using the OpenSSH sshd
implementation.  It turns out that - with the right configuration - sshd
is perfectly capable of running without privileges.  Also, use a custom
ssh key for authentication to not rely at all on the users ssh
configuration.

Use this new mock-SSH architecture for all tests which should run
against all linux shell implementations to get good coverage.

If OpenSSH is not installed, skip all SSH-related tests.

Signed-off-by: Harald Seiler <hws@denx.de>