path: Actually handle errors in write_text() and write_binary()
Currently, `Path.write_text()` and `Path.write_binary()` configure
a death-string but never act upon it being triggered.  This would not
show up in most cases because the death-string was never seen on fast
channels.

On slow channels, however, it does trigger.  This means we need to
properly handle it.  In this case, this just means aborting early with
the same sequence of sending `^D` and waiting for the command to
complete.

Signed-off-by: Harald Seiler <hws@denx.de>