treewide: Invoke bash with --noediting
With GNU readline 8.1, bracketed paste is enabled by default which tbot
cannot currently handle.  This change showed that we still have some
places in tbot where readline was (unintentionally) not disabled.  Fix
this by adding `--noediting` to each shell invocation to ensure we do
not have the GNU readline active.

Ref: #51
Signed-off-by: Harald Seiler <hws@denx.de>