linux.path: Drop pathlib inheritance
It was a mistake to make tbot's Path inherit from pathlib:  While both
provide similar interfaces (intentionally), a tbot-path is **not**
supposed to be usable as a pathlib-path, as it represents
a fundamentally different thing.

Drop the inheritance relationship and instead just rely on pathlib
internally for managing the path juggling.

Additionally, drop the __new__() method from tbot's GitRepository class
as it was only needed due to inheritance weirdness.

Signed-off-by: Harald Seiler <hws@denx.de>