linux.path: Add method to convert Path to str safely
A longstanding problem with tbot's Path was that there was no clean way
to convert it to a string.  It was designed like this because in most
situations, conversion to a string is the wrong thing to do.  However,
there are some valid uses for this and we need to provide a method to
allow them.

`at_host` is this method:  It asserts that the path is associated with
the given host and if yes, it returns the string representation.  If the
host does not match, an exception is raised.

This design ensures that a programmer using the method has to at least
think for a second about which host they are trying to use the path on
and this hopefully catches most stupid mistakes...

Ref: #53
Co-authored-by: Bernhard Kirchen <bernhard.kirchen@mbconnectline.com>
Signed-off-by: Harald Seiler <hws@denx.de>