context: Allow reconfiguring a context temporarily
Add a new method to the `Context` to open a temporary reconfiguration
where flags like `keep_alive` and `reset_on_error_by_default` can be
updated:

    with ctx.reconfigure(keep_alive=True):
        ...

Once the reconfiguration ends, the old state is restored.  This also
means that any instances which were kept-alive due to the
reconfiguration will now be torn down.

Signed-off-by: Harald Seiler <hws@denx.de>