board: Add powercycle_delay to PowerControl initializer
Add a `powercycle_delay` attribute to the `PowerControl` initializer
which can be used to specify how much time tbot needs to wait between
powering off the board and powering it back on.

The delay is only performed when a poweroff is followed by a poweron,
and thus avoids unnecessary time waste when e.g. powering on in the
beginning and powering off at the end of a testrun.

Signed-off-by: Harald Seiler <hws@denx.de>