newbot: Add initial version
`newbot` is going to be the new CLI for tbot.  For now, it will be
developed as a separate tool as this allows experimentation with
breaking changes without breaking existing code.  In the long run this
new CLI will replace the `tbot` tool (and assume its name, of course).

Why?
----
The reason for this new CLI is mostly that the old one has a number of
rather ugly problems.  Addressing these problems in place is difficult -
it is much easier to rework the CLI entirely to provide the best
interface going forward.  The issues that I try to solve are:

1. The old configuration mechanism is quite rigid and often does not map
   well to usage patterns.  Especially the requirement for exactly two
   configurations (a lab and a board) might match some peoples reality
   but certainly isn't the most common deployment.
2. The old configuration and testcase loading mechanism is using paths
   to Python scripts.  This meant that a custom script loader was
   required with did not mesh well with the python import system at all.
   Specifically, there was no way to reference a configuration module
   from testcase code.  There were also inconsistencies where certain
   code works when written at module level but doesn't when inside a
   function called later on.
3. Building parameterized configuration was super difficult and often
   required modifications to the tbot CLI utility.  Unfortunately, this
   is a common requirement and as such should be well supported.
4. The notion of testcases by use of a magic decorator was always a bit
   iffy.  Again, there were problems with making this mesh well with
   Python's module mechanism.

To address all these, the concept of configuration and testcase modules
is completely redesigned with `newbot`.  Config and testcases are now
referenced by Python module path and as such completely rely on said
mechanism instead of our old homebuilt version.  Additionally, any
notion of types of config (lab-config, board-config) is removed -
instead, any config can register whatever machines it wants.  This
allows the most flexibility in splitting or not splitting the configs.

Finally, a last goal of `newbot` is to make the CLI as minimal and
modular as possible.  tbot should not be a framework dictating how one
should use it but instead be a libary for console interaction.  The CLI
is then just a "cherry on top" for easily invoking scripts built with
this library with a nice UX and TUI.

Signed-off-by: Harald Seiler <hws@denx.de>