tc: uboot: Make uboot_testpy honor `powercycle_delay`
When the chosen board has configured a `powercycle_delay` [1] as
introduced in commit ded6c614c338 ("board: Add powercycle_delay to
PowerControl initializer") the `uboot_testpy` testcase currently
misbehaves as this delay is not performed between poweroff and poweron
of the board.

Fix this by adding a sleep for the configured duration if the attribute
is set.

[1]: https://tbot.tools/modules/machine_board.html#tbot.machine.board.PowerControl.powercycle_delay
Signed-off-by: Harald Seiler <hws@denx.de>