newbot: Add --keep-alive option
Add a --keep-alive option to newbot which enables the keep-alive feature
of `tbot.Context`.  This allows running testcases in series faster as a
machine will be instantly available to later testcases after it was
initialized once by the first one.

You can still reboot a machine between testcases by writing a "reboot"
testcase which uses `Context.teardown_if_alive()` and is called in
between the two other testcases.

Signed-off-by: Harald Seiler <hws@denx.de>