context: Add teardown_if_alive() method
The teardown_if_alive() method allows to ensure that no instance of a
certain role is active after it is called.  This can be useful, for
example, when a `BoardLinux` instance is alive but you need to enter
U-Boot.  Normally this would fail, because the `BoardLinux` instance is
keeping the `Board` instance for itself.  With teardown_if_alive() you
can first tear down this `BoardLinux` instance and then request the
`BoardUBoot` instance as usual.

Signed-off-by: Harald Seiler <hws@denx.de>