context: Don't count pytest.skip() for reset_on_error
When reset_on_error is enabled and an exception is raised, instances in
the context are reset.  This is useful to bring them back to a known
state after an error exception was raised, but is annoying when
exceptions are thrown for other reasons.  For example, the
`pytest.skip()` function uses an exception to exit a testcase
prematurely.

Detect this special exception in tbot's context and ignore
`reset_on_error` when it is raised to keep instances alive when tests
are skipped.

Signed-off-by: Harald Seiler <hws@denx.de>