contrib: locking: Disable reset_on_error for labhost access
When `reset_on_error_by_default` is enabled in a Context, the "could not
acquire lock exceptions" thrown here will lead to the labhost getting
torn down.  This is not really what is intended and can lead to further
oddities down the line.

Explicitly disable `reset_on_error` for the requests done in MachineLock
implementations to prevent this situation.

Signed-off-by: Harald Seiler <hws@denx.de>