channel: Use individual ringbuffers per death-string
Using a single ringbuffers has a large problem: When a new deathstring
is added, it will potentially get checked against bytes which were
received prior to the string being added.  This bug surfaces, for
example, when adding a long death-string to a machine and then calling
`LinuxShell.run()` on it.

Fix this by switching to individual ringbuffers per death-string.  Thus
each string is only checked against bytes relevant to it - and at the
same time the logic gets a bit more compact because we can drop the
length management magic...

Also add a testcase to verify this behavior won't regress in the future.

Signed-off-by: Harald Seiler <hws@denx.de>