linux.path: Fix write_bytes()/write_test() raising wrong exception
By mistake, the ch.terminate0() call was placed inside the with-block
for the tee death-string.  This means it might potentially re-raise the
exception for this death-string (as it is still active here) which leads
to an improper exception being propagated to the caller.  It also
potentially means that the shell invocations on the other side are not
cleaned up properly, leading to strange errors further down the line...

Fix this by moving the ch.terminate0() call outside of the with-block
which means the death-string is no longer active and the run-command can
be finished properly.

Signed-off-by: Harald Seiler <hws@denx.de>