selftest: Ensure only intended exceptions are caught
Unfortunately, tbot does not yet raise distinct exceptions for each
"error type".  Thus, lot's of tests just checked for _any_ exception.

This is problematic because it hides cases where the wrong exception is
raised by a testcase.  Thus, to be robust against such bugs, always
match on the exception in an unambiguous way.

Signed-off-by: Harald Seiler <hws@denx.de>