contrib: uboot: Add reworked test/py integration
There exists previous integration for U-Boot's test/py testsuite in
`tbot.tc.uboot`.  This implementation, however, has bitrotten quite some
and does not work well with newer versions of tbot anymore - for some
setups it fails to work at all.

The problem was that this old testcase aimed to be directly callable and
for this it needs to figure out its environment on its own:  Where
U-Boot sources are found, what U-Boot config to use, what board machine
to use, etc.  This amount of "smart detection" made the testcase often
too inflexible for adequate use.

Add a reworked test/py integration which instead provides only the
mechanism to attach a tbot board-machine to a test/py invocation from an
existing U-Boot source tree.  This mechanism should then be integrated
into ones own calling testcase.  As an example, this might be what such
calling code could look like:

    from tbot_contrib import uboot

    BOARDENV = r"""# Boardenv for xyz board.

    env__net_dhcp_server = True
    """

    @tbot.testcase
    def run_testpy() -> None:
        with tbot.ctx.request(tbot.role.BuildHost) as h:
            # location of the U-Boot sources - these must have been
            # pre-configured elsewhere (or configured here manually)
            build_dir = h.workdir / "u-boot-mainline"

            # subshell for the build environment
            with h.subshell():
                # setup toolchain
                h.env("CROSS_COMPILE", "arm-linux-")

                # if needed, you could configure sources here
                # h.exec0("make", "xyz_defconfig")

                uboot.testpy(
                    build_dir,
                    boardenv=BOARDENV,
                    testpy_args=["--maxfail", "6"],
                )

If the default of connecting to the selected board-machine is not
correct, you could also pass board and U-Boot machines manually:

    @tbot.testcase
    def run_testpy() -> None:
        with tbot.ctx() as cx:
            h = cx.request(tbot.role.BuildHost)
            build_dir = ...
            ...

            # for demonstration - this is exactly what uboot.testpy()
            # would do on its own when board and uboot are not passed.
            b = cx.request(tbot.role.Board)
            ub = cx.request(tbot.role.BoardUBoot, exclusive=True)

            uboot.testpy(
                build_dir,
                board=b,
                uboot=ub,
                boardenv=BOARDENV,
            )

It is advisable to request the U-Boot machine with exclusive=True so no
other code will attempt interacting with it while testpy is running.
This is not a hard requirement, though.

Signed-off-by: Harald Seiler <hws@denx.de>