tc: Deprecate old selftest suite
The old selftest suite (invoked with `tbot selftest`) has not been
maintained in quite some time.  Let's deprecate it for good and make it
throw an error instead of attempting to run the old tests.

You can still force the tests with `-f really-run-old-selftests` if you
for some reason need them.

Signed-off-by: Harald Seiler <hws@denx.de>