linux.path: Implement Path.resolve()
`pathlib` provides a `resolve()` method to turn a relative link absolute
and resolve any symlink components.  Implement this method for tbot
paths as well.

Due to using `realpath(1)` underneath, the behavior of `strict=False`
differs slightly:  Only the last component is allowed to be missing
instead of an arbitrarily long suffix.

Signed-off-by: Harald Seiler <hws@denx.de>