linux.path: Improve glob() implementation
Fix `Path.glob()` not escaping the glob-pattern.  This means switching
to `find` underneath instead of relying on the shell for globbing.  This
is not ideal but at least ensures a consistent experience across
implementations (to some degree).

Important to note, this will remove the ability to glob in multiple path
components at the same time.  Only the last path component may contain
globbing.

In addition, improve the testsuite to cover odd filenames as well and
use pytest parameterization for better reporting.

Signed-off-by: Harald Seiler <hws@denx.de>