Update python packaging to newer standards
The world has moved on and now a different set of magic files are needed
for Python packaging.  Update to the current standard and move as far
away from the legacy setup.py as we can.

In the same move, switch to setuptools-scm which automatically retrieves
the project version from git.  This will provide much more consistent
version reporting...

Signed-off-by: Harald Seiler <hws@denx.de>