channel: subprocess: Fix non-blocking writes
As we configure the pty_master file-descriptor for O_NONBLOCK, the
writer code must also be set-up to correctly handle non-blocking I/O.

Do this by first calling select() to block until the file-descriptor
becomes writeable.  If it does not get there in 10 seconds, abort with a
timeout error as this is really unusual.

Signed-off-by: Harald Seiler <hws@denx.de>