channel: Improve behavior of send() for large amounts of data
When calling send() with large amounts of data and having `read_back`
enabled, it would first try to get all data sent before doing the
read-back.  This only works as long as the buffer behind the PTY is
large enough.

Fix this by sending data in chunks of maximum 512 bytes.  These 512
bytes will have to be read back first before the next chunk is sent.

Signed-off-by: Harald Seiler <hws@denx.de>