board.linux: Allow logging in without password if not asked
Sometimes, a system is used with different Linux rootfs images, some of
which ask for a password and some don't.  Up to now, this had to be
handled by having different configurations for each.

This is a bit annoying.  Add a mechanism which allows tbot to continue
without entering the password when no password prompt shows up in a
given time.  Setting `password = None` should still be preferred, but
this makes tbot a bit more forgiving in that regard.

This mechanism has one drawback:  If the system gets stuck right after
tbot entered the username and then never reaches the password prompt,
the `boot_timeout` of the machine will not be honored anymore.  Instead,
tbot will search for a shell response indefinitely.  If you are
concerned about this, set `no_password_timeout = None` to be certain
about the timeout applying.

Signed-off-by: Harald Seiler <hws@denx.de>